T3.1-dynamic-analysis/Case-study-I-EPCs-classification/Code/requirements.txt
affine==2.3.1
anyio==3.6.2
appdirs==1.4.4
appnope==0.1.3
arcgis==1.9.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asciitree==0.3.3
asttokens==2.0.5
attrs==21.4.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
bokeh==2.4.2
Bottleneck==1.3.5
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.0
cfgrib==0.9.10.3
cftime==1.6.2
charset-normalizer==2.1.0
click==8.0.4
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.0.0
cryptography==37.0.1
cycler==0.11.0
Cython==0.29.32
cytoolz==0.11.0
dask==2022.5.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
distributed==2022.5.0
eccodes==1.5.0
ecmwflibs==0.5.0
entrypoints==0.4
executing==0.8.3
fasteners==0.18
fastjsonschema==2.16.2
findlibs==0.0.2
Fiona==1.8.21
flox==0.6.4
fonttools==4.34.4
fsspec==2022.3.0
future==0.18.2
geocoder==1.38.1
geographiclib==1.52
geomet==1.0.0
geopandas==0.11.0
geopy==2.2.0
gssapi==1.8.2
h5netcdf==1.1.0
h5py==3.7.0
HeapDict==1.0.1
idna==3.3
ipykernel==6.9.1
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
json5==0.9.10
jsonschema==4.17.3
jupyter-client==7.2.2
jupyter-core==4.10.0
jupyter-server==1.23.3
jupyterlab==3.5.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
jupyterlab_server==2.16.3
keyring==21.7.0
kiwisolver==1.4.3
lerc==0.1.0
llvmlite==0.39.1
locket==1.0.0
lxml==4.9.1
lz4==3.1.3
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.2
mistune==2.0.4
-e git+ssh://git@gitlab.inf.unibz.it/URS/moderate.git@05164e06b2ac5db61ffbb9c0daa67b524d1a354f#egg=moderate
msgpack==1.0.3
munch==2.5.0
nbclassic==0.4.8
nbclient==0.6.3
nbconvert==7.2.5
nbformat==5.7.0
nc-time-axis==1.4.1
nest-asyncio==1.5.5
netCDF4==1.6.2
networkx==2.8.8
notebook==6.5.2
notebook_shim==0.2.2
ntlm-auth==1.5.0
numba==0.56.4
numbagg==0.2.1
numcodecs==0.10.2
numexpr==2.8.3
numpy==1.23.0
numpy-groupies==0.9.20
oauthlib==3.2.2
openeo==0.10.1
osmnx==1.2.2
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
partd==1.2.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pooch==1.6.0
prometheus-client==0.15.0
prompt-toolkit==3.0.20
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
-e git+https://git@github.com/aaronestrada/py-config-env.git@3e67397118bc5eef2c0acc3c772fa60741728a63#egg=py_config_env
pycparser==2.21
Pygments==2.11.2
pygrib==2.1.4
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyproj==3.3.1
pyrsistent==0.19.2
pyshp==2.3.1
PySocks==1.7.1
python-certifi-win32==1.6.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
pyzmq==22.3.0
rasterio==1.3.2
rasterstats==0.17.0
ratelim==0.1.6
requests==2.28.1
requests-gssapi==1.2.3
requests-ntlm==1.1.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
Rtree==1.0.1
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.1
Send2Trash==1.8.0
setuptools-scm==7.0.5
Shapely==1.8.2
simplejson==3.18.0
six==1.16.0
sklearn==0.0
sniffio==1.3.0
snuggs==1.4.7
sodapy==2.1.1
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.2.0
tblib==1.7.0
terminado==0.17.0
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
toolz==0.11.2
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typing_extensions==4.1.1
ujson==5.5.0
Unidecode==1.3.6
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.14.1
xarray==2022.11.0
xgboost==1.7.1
yellowbrick==1.5
zarr==2.13.3
zict==2.1.0
